fix: id 필드 타입 수정 및 일부 의존성 추가

See merge request s13-bigdata-dist-sub1/S13P21A302!12

+++ backend-spring/leaper/src/main/java/com/ssafy/leaper/domain/insight/dto/response/contentCommentInsight/ContentCommentInsightResponse.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 public record ContentCommentInsightResponse(
-    Long contentId,
+    Integer contentId,
     String contentTitle,
     List<String> positiveComments,
     List<String> negativeComments,
